Split Effect thunk fixtures
Amp-Thread-ID: https://ampcode.com/threads/T-01a03991-b94f-76da-ada4-134b67f32b93
