+++ tests-integration/fixtures/backend/1787673600_effect_ado_thunks/output/Main/index.js
@@ -14,38 +14,4 @@ export function timedAdo(seed) {
   return $effect;
 }
 
-export function identity(value) {
-  return value;
-}
-
-export function mapped(value) {
-  const $function = mark("map-function")(identity);
-  const $action = constructEffect("map-action")(value);
-  const $effect = () => {
-    const $value = $action();
-    return $function($value);
-  };
-  return $effect;
-}
-
-export function applied(value) {
-  const $functionAction = constructEffect("apply-function-action")(identity);
-  const $argumentAction = constructEffect("apply-value-action")(value);
-  const $effect = () => {
-    const $function = $functionAction();
-    const $argument = $argumentAction();
-    return $function($argument);
-  };
-  return $effect;
-}
-
-export function capturedPure(value) {
-  const $value = mark("pure-value")(value);
-  const $effect = () => {
-    return $value;
-  };
-  return $effect;
-}
-
 export const constructEffect = $foreign["constructEffect"];
-export const mark = $foreign["mark"];
