+++ tests-integration/fixtures/backend/1787673600_effect_ado_thunks/verify.mjs
@@ -8,39 +8,10 @@ const timedAdoConstruction = readTrace();
 const timedAdoValue = timedAdo();
 const timedAdoTrace = readTrace();
 
-resetTrace();
-const mapped = Main.mapped("mapped");
-const mappedConstruction = readTrace();
-const mappedValue = mapped();
-const mappedTrace = readTrace();
-
-resetTrace();
-const applied = Main.applied("applied");
-const appliedConstruction = readTrace();
-const appliedValue = applied();
-const appliedTrace = readTrace();
-
-resetTrace();
-const capturedPure = Main.capturedPure("pure");
-const capturedPureConstruction = readTrace();
-const capturedPureFirst = capturedPure();
-const capturedPureSecond = capturedPure();
-const capturedPureTrace = readTrace();
-
 const actual = {
   timedAdoConstruction,
   timedAdoValue,
   timedAdoTrace,
-  mappedConstruction,
-  mappedValue,
-  mappedTrace,
-  appliedConstruction,
-  appliedValue,
-  appliedTrace,
-  capturedPureConstruction,
-  capturedPureFirst,
-  capturedPureSecond,
-  capturedPureTrace,
 };
 
 const expected = {
@@ -52,28 +23,6 @@ const expected = {
     "run:ado-first",
     "run:ado-second",
   ],
-  mappedConstruction: ["mark:map-function", "construct:map-action"],
-  mappedValue: "mapped",
-  mappedTrace: [
-    "mark:map-function",
-    "construct:map-action",
-    "run:map-action",
-  ],
-  appliedConstruction: [
-    "construct:apply-function-action",
-    "construct:apply-value-action",
-  ],
-  appliedValue: "applied",
-  appliedTrace: [
-    "construct:apply-function-action",
-    "construct:apply-value-action",
-    "run:apply-function-action",
-    "run:apply-value-action",
-  ],
-  capturedPureConstruction: ["mark:pure-value"],
-  capturedPureFirst: "pure",
-  capturedPureSecond: "pure",
-  capturedPureTrace: ["mark:pure-value"],
 };
 
 deepStrictEqual(actual, expected);
