+++ tests-integration/fixtures/backend/1787673600_effect_do_thunks/output/Main/index.js
@@ -1,6 +1,4 @@
-import * as Control_Bind from "../Control.Bind/index.js";
 import * as Data_Unit from "../Data.Unit/index.js";
-import * as Effect from "../Effect/index.js";
 import * as $foreign from "./foreign.js";
 
 export function chained(seed) {
@@ -34,52 +32,5 @@ export function pureAfterBind(seed) {
   return $effect;
 }
 
-export function branched(choose) {
-  return seed => {
-    const $action = constructEffect("branch-action")(seed);
-    const $effect = () => {
-      const value = $action();
-      if (choose) {
-        return constructEffect("branch-then")(value)();
-      } else {
-        return constructEffect("branch-else")(value)();
-      }
-    };
-    return $effect;
-  };
-}
-
-export function patternLet(seed) {
-  const $action = constructEffect("pattern-action")(seed);
-  const $effect = () => {
-    const value = $action();
-    const $scrutinee = { selected: value };
-    const selected = $scrutinee.selected;
-    return constructEffect("pattern-result")(selected)();
-  };
-  return $effect;
-}
-
-export function genericBind(bindMDict) {
-  return Control_Bind.bind(bindMDict);
-}
-
-export function aliased(seed) {
-  return genericBind(Effect.bindEffect)(constructEffect("alias-first")(seed))(
-    value => constructEffect("alias-second")(value)
-  );
-}
-
 export const constructEffect = $foreign["constructEffect"];
 export const mark = $foreign["mark"];
-
-export const deferredEffect = (() => {
-  const $action = constructEffect("deferred-action")("ignored");
-  const $effect = () => {
-    const value = $action();
-    return constructEffect("deferred-result")(deferredValue)();
-  };
-  return $effect;
-})();
-
-export const deferredValue = mark("deferred-value")("deferred");
