+++ tests-integration/fixtures/backend/1787673600_effect_do_thunks/verify.mjs
@@ -2,8 +2,6 @@ import { deepStrictEqual } from "node:assert/strict";
 import * as Main from "./output/Main/index.js";
 import { readTrace, resetTrace } from "./output/Main/foreign.js";
 
-const startupTrace = readTrace();
-
 resetTrace();
 const chained = Main.chained("seed");
 const chainedConstruction = readTrace();
@@ -24,32 +22,7 @@ const pureAfterBindConstruction = readTrace();
 const pureAfterBindValue = pureAfterBind();
 const pureAfterBindTrace = readTrace();
 
-resetTrace();
-const branchedThen = Main.branched(true)("then");
-const branchedThenConstruction = readTrace();
-const branchedThenValue = branchedThen();
-const branchedThenTrace = readTrace();
-
-resetTrace();
-const branchedElse = Main.branched(false)("else");
-const branchedElseConstruction = readTrace();
-const branchedElseValue = branchedElse();
-const branchedElseTrace = readTrace();
-
-resetTrace();
-const patternLet = Main.patternLet("pattern");
-const patternLetConstruction = readTrace();
-const patternLetValue = patternLet();
-const patternLetTrace = readTrace();
-
-resetTrace();
-const aliased = Main.aliased("alias");
-const aliasedConstruction = readTrace();
-const aliasedValue = aliased();
-const aliasedTrace = readTrace();
-
 const actual = {
-  startupTrace,
   chainedConstruction,
   chainedFirst,
   chainedFirstTrace,
@@ -61,22 +34,9 @@ const actual = {
   pureAfterBindConstruction,
   pureAfterBindValue,
   pureAfterBindTrace,
-  branchedThenConstruction,
-  branchedThenValue,
-  branchedThenTrace,
-  branchedElseConstruction,
-  branchedElseValue,
-  branchedElseTrace,
-  patternLetConstruction,
-  patternLetValue,
-  patternLetTrace,
-  aliasedConstruction,
-  aliasedValue,
-  aliasedTrace,
 };
 
 const expected = {
-  startupTrace: ["construct:deferred-action", "mark:deferred-value"],
   chainedConstruction: ["construct:first"],
   chainedFirst: { first: "seed", second: { first: "seed" } },
   chainedFirstTrace: [
@@ -116,38 +76,6 @@ const expected = {
     "construct:pure-action",
     "run:pure-action",
     "mark:pure-body",
-  ],
-  branchedThenConstruction: ["construct:branch-action"],
-  branchedThenValue: "then",
-  branchedThenTrace: [
-    "construct:branch-action",
-    "run:branch-action",
-    "construct:branch-then",
-    "run:branch-then",
-  ],
-  branchedElseConstruction: ["construct:branch-action"],
-  branchedElseValue: "else",
-  branchedElseTrace: [
-    "construct:branch-action",
-    "run:branch-action",
-    "construct:branch-else",
-    "run:branch-else",
-  ],
-  patternLetConstruction: ["construct:pattern-action"],
-  patternLetValue: "pattern",
-  patternLetTrace: [
-    "construct:pattern-action",
-    "run:pattern-action",
-    "construct:pattern-result",
-    "run:pattern-result",
-  ],
-  aliasedConstruction: ["construct:alias-first"],
-  aliasedValue: "alias",
-  aliasedTrace: [
-    "construct:alias-first",
-    "run:alias-first",
-    "construct:alias-second",
-    "run:alias-second",
   ],
 };
