+++ tests-integration/fixtures/backend/1787678880_effect_do_boundaries/output/Effect/foreign.js
@@ -1,0 +1,4 @@
+export const mapE = transform => action => () => transform(action());
+export const applyE = functionAction => valueAction => () => functionAction()(valueAction());
+export const pureE = value => () => value;
+export const bindE = action => continuation => () => continuation(action())();
