+++ tests-integration/fixtures/backend/1787678880_effect_do_boundaries/output/Main/index.js
@@ -1,0 +1,27 @@
+import * as Control_Bind from "../Control.Bind/index.js";
+import * as Effect from "../Effect/index.js";
+import * as $foreign from "./foreign.js";
+
+export function genericBind(bindMDict) {
+  return Control_Bind.bind(bindMDict);
+}
+
+export function aliased(seed) {
+  return genericBind(Effect.bindEffect)(constructEffect("alias-first")(seed))(
+    value => constructEffect("alias-second")(value)
+  );
+}
+
+export const constructEffect = $foreign["constructEffect"];
+export const mark = $foreign["mark"];
+
+export const deferredEffect = (() => {
+  const $action = constructEffect("deferred-action")("ignored");
+  const $effect = () => {
+    const value = $action();
+    return constructEffect("deferred-result")(deferredValue)();
+  };
+  return $effect;
+})();
+
+export const deferredValue = mark("deferred-value")("deferred");
