+++ tests-integration/fixtures/backend/1787678880_effect_do_control_flow/Main.js
@@ -1,0 +1,15 @@
+const trace = [];
+
+export const resetTrace = () => {
+  trace.length = 0;
+};
+
+export const readTrace = () => trace.slice();
+
+export const constructEffect = label => value => {
+  trace.push(`construct:${label}`);
+  return () => {
+    trace.push(`run:${label}`);
+    return value;
+  };
+};
