+++ tests-integration/fixtures/backend/1787678880_effect_do_control_flow/output/Control.Bind/index.js
@@ -1,0 +1,9 @@
+export function bind(dictionary) {
+  return dictionary.bind;
+}
+
+export function discard(dictionary) {
+  return dictionary.discard;
+}
+
+export const discardUnit = { discard: bindFDict => bind(bindFDict) };
