+++ tests-integration/fixtures/backend/1787678880_effect_do_control_flow/package.json
@@ -1,0 +1,1 @@
+{"type":"module"}
