+++ tests-integration/fixtures/backend/1787678880_effect_primitive_thunks/Main.js
@@ -1,0 +1,20 @@
+const trace = [];
+
+export const resetTrace = () => {
+  trace.length = 0;
+};
+
+export const readTrace = () => trace.slice();
+
+export const constructEffect = label => value => {
+  trace.push(`construct:${label}`);
+  return () => {
+    trace.push(`run:${label}`);
+    return value;
+  };
+};
+
+export const mark = label => value => {
+  trace.push(`mark:${label}`);
+  return value;
+};
