+++ tests-integration/fixtures/backend/1787678880_effect_primitive_thunks/output/Control.Applicative/index.js
