+++ tests-integration/fixtures/backend/1787678880_effect_primitive_thunks/output/Control.Apply/index.js
@@ -1,0 +1,7 @@
+import * as Data_Functor from "../Data.Functor/index.js";
+
+export function apply(dictionary) {
+  return dictionary.apply;
+}
+
+export const applyFn = { Functor0: () => Data_Functor.functorFn, apply: f => g => x => f(x)(g(x)) };
