+++ tests-integration/fixtures/backend/1787678880_effect_primitive_thunks/output/Data.Functor/index.js
@@ -1,0 +1,5 @@
+export function map(dictionary) {
+  return dictionary.map;
+}
+
+export const functorFn = { map: f => g => x => f(g(x)) };
