+++ tests-integration/fixtures/backend/1787678880_effect_primitive_thunks/package.json
@@ -1,0 +1,1 @@
+{"type":"module"}
